According to godbolt. It generated better code.
Also the 2 benches I run shows a improvement with the `email_parser` test.


Original code:
```
running 4 tests
test email        ... bench:      17,097 ns/iter (+/- 3,712)
test email_format ... bench:     403,548 ns/iter (+/- 26,380)
test email_parser ... bench:      75,946 ns/iter (+/- 20,761)
test mailparse    ... bench:       6,329 ns/iter (+/- 368)
```

My improvement:
```
running 4 tests
test email        ... bench:      17,911 ns/iter (+/- 3,563)
test email_format ... bench:     433,603 ns/iter (+/- 82,618)
test email_parser ... bench:      45,976 ns/iter (+/- 4,599)
test mailparse    ... bench:       6,727 ns/iter (+/- 869)
```