@vDorst 2020-09-01
@Mubelotix No special email.
Just picked one from my mailbox to test.

ls -al mail.txt
```
42004 bytes mail.txt
```
Was an email from Sony Playstation Network.